+++ R/tar_map.R
@@ -72,10 +72,11 @@ tar_map <- function(
   unlist = FALSE
 ) {
   targets <- unlist(list(...), recursive = TRUE)
-  assert_targets(targets)
+  targets::tar_assert_target_list(targets)
   assert_values_list(values)
   names_quosure <- rlang::enquo(names)
   names <- eval_tidyselect(names_quosure, base::names(values))
+  values <- tibble::as_tibble(values)
   values <- tar_map_process_values(values)
   values <- tar_map_extend_values(targets, values, names)
   out <- lapply(targets, tar_map_target, values = values)
@@ -91,7 +92,7 @@ tar_map_process_values <- function(values) {
   for (name in names(values)) {
     values[[name]] <- map(
       values[[name]],
-      ~parse(text = deparse_safe(.x))[[1]]
+      ~parse(text = targets::tar_deparse_safe(.x))[[1]]
     )
   }
   values
@@ -101,7 +102,7 @@ tar_map_extend_values <- function(targets, values, names) {
   suffix <- tar_map_produce_suffix(values, names)
   for (target in targets) {
     name <- target$settings$name
-    assert_not_in(
+    targets::tar_assert_not_in(
       name,
       names(values),
       paste("target", name, "cannot be in names(values).")
@@ -112,12 +113,41 @@ tar_map_extend_values <- function(targets, values, names) {
 }
 
 tar_map_produce_suffix <- function(values, names) {
-  data <- values[names] %||% list(id = seq_along(values[[1]]))
+  data <- values[names] %||% tar_map_default_suffixes(values)
   data <- map(data, ~as.character(unlist(.x)))
   out <- apply(as.data.frame(data), 1, paste, collapse = "_")
   out <- gsub("'", "", out)
   out <- gsub("\"", "", out)
   make.unique(out, sep = "_")
+}
+
+tar_map_default_suffixes <- function(values) {
+  # TODO: remove message in the next release
+  message <- paste0(
+    "INVALIDATING CHANGE: in tarchetypes version 0.3.0 and above, ",
+    "tar_map() branch names are now hashes instead of numeric indexes ",
+    "whenever the `names` argument is NULL. This helps pipelines become ",
+    "more stable in the long run because reordring rows of `values` ",
+    "no longer invalidates targets by changing the target names. However, ",
+    "if this is your first time upgrading to tarchetypes 0.3.0 or above, ",
+    "then some of your targets may be invalidated once in the next run of ",
+    "your pipeline. For details, please visit ",
+    "https://github.com/ropensci/tarchetypes/issues/67. ",
+    "Sorry for the inconvenience."
+  )
+  rlang::inform(
+    message,
+    class = "tar_condition_validate",
+    .frequency = "regularly",
+    .frequency_id = "tarchetypes_67"
+  )
+  id <- apply(
+    X = values,
+    MARGIN = 1,
+    FUN = digest::digest,
+    algo = "xxhash32"
+  )
+  list(id = id)
 }
 
 tar_map_target <- function(target, values) {